Created a logo-upload client-side controller.
